Update CDASimAdapter to accomadate changes in SensorConfig file (#640)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
### Background (Related changes)
This PR was created to accommodate changes in the Sensor Configuration
file introduced by this PR
(https://github.com/usdot-fhwa-stol/carma-streets/pull/427). To
summarize, the sensor configuration file format was changed to allow for
reference location to be specified in WGS84 coordinates or cartesian
coordinates. For Simulation cartesian coordinates work well since many
simulation environments do not have real-world equivalents and cartesian
coordinates are easy to pull from different map editors and viewers like
[JSOM](https://josm.openstreetmap.de/). For real world deployments
however WGS84 coordinates are far easier to use since using cartesian
coordinates introduces a dependency on an lanlet2 map for the area. It
is far easier to use built in or separate GPS devices to get a reference
location for a mounted sensor.
#### Old Format of Sensor Configuration JSON:
```json
[
  {
    "sensorId": "SomeID",
    "type": "SemanticLidar",
    "location": {
      "x": 1.0, /* in meters */
      "y": 2.0, /* in meters */
      "z": -3.2 /* in meters */
     },
     "orientation": {
       "yaw": 0.0,
       "pitch": 0.0,
       "roll": 0.0
      }
  }
]
```  
#### New Format of Sensor Configuration JSON:
```json
[
  {
    "sensorId": "sensor_1",
    "type": "SemanticLidar",
    "ref": {
      "type": "CARTESIAN",
      "location": {
        "x": 1.0, /* in meters */
        "y": 2.0, /* in meters */
        "z": -3.2 /* in meters */
      },
       "orientation": { 
        "yaw": 0.0,
        "pitch": 0.0,
        "roll": 0.0
        }
    }
  } 
]
```

### Changes

This change in Sensor Configuration file means that we need to change
how we read and populate the registration message that goes out from the
CDASim Adapter Plugin to CDASim. The content of this registration
remains unchanged, we just need to read this information from the new
`ref` element in the Sensor Configuration JSON and populate the existing
registration. Lastly, we added logic for the registration to ignore any
sensor with a WGS84 reference location since we currently only have
support
```json
{
  "infrastructureId": "id_123",
  "location": {
    "x": 0,
    "y": 0,
    "z": 0
  },
  "rxMessageIpAddress": "127.0.0.1",
  "rxMessagePort": 8686,
  "sensors": {
    "sensorId": "sensor1",
    "type": "SemanticLidar",
    "location": {
      "x": 0.0,
      "y": 0.0,
      "z": 0.0
     },
     "orientation": {
       "yaw": 0.0,
       "pitch": 0.0,
       "roll": 0.0
      }
  },
  "simulatedInteractionPort": 7576,
  "timeSyncPort": 7575
}

```

<!--- Describe your changes in detail -->

## Related Issue
[FCP-37
](https://usdot-carma.atlassian.net/browse/FCP-37)<!--- This project
only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
See description
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Local Integration Testing and unit testing
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.