Arc 239 asn1 j2735 (#651)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
First reverted back to the ASN1 generated scripts. After doing so, names
of enums in the IVPJ2735.h and .c needed to be changed due to name
conflicts with the macros defined in the generated structs. A few
pointers needed to be fixed in the J2735MessageTest file as well as more
naming issues in the Csw and Tim Plugins.
<!--- Describe your changes in detail -->

## Related Issue
https://usdot-carma.atlassian.net/browse/ARC-239
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
Currently V2X Hub has manual changes to the ASN1C Compiler generator
code that make it incompatible with our carma-j2735-asn package. This
story is to capture efforts of undoing those manual changes both in the
generated code and in the logic we use to populate/read from generated C
Structs.
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
To test these changes, a full successful build needed to be completed
without any errors being thrown. Currently, the updated V2X Hub is able
to successfully build, but there are some unit tests that need to be
fixed.


<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.

---------

Co-authored-by: root <root@Ubuntu.myguest.virtualbox.org>
Co-authored-by: Satish Natarajan <natarajan.sa@northeastern.edu>