MOM and ActionObject Converters, Miscellaneous MOM handling updates for CDA 1Tenth (#674)
<!-- Thanks for the contribution, this is awesome. -->

Modifications to the MOM handling of the Port Drayage Plugin were needed
to adapt them to the CDA 1Tenth plugin. The major change required was
updating the CDA 1Tenth Plugin to support the newly defined "Action
Object", replacing the previous "Port Drayage Object." Files have been
separated based on functionality to improve readability and modularity.

NOTE 1: Some changes related to the database and action representation
are NOT updated since this may not be finalized. Fixes to be addressed
due to this reason are marked with "TODO"

NOTE 2: Some BSM and Websocket changes have NOT been merged into this
branch yet so that changes are easier to track and review. When the
draft PR is turned into the final PR this branch will be updated before
final review.

<!--- Describe your changes in detail -->

CF-969

Data representation enhancements, handling data for MOMs.

Unit tests written and passed for this plugin. Unknown Kafka test core
dump occurs when running all unit tests but almost certainly not related
to this PR. See below.

![image](https://github.com/user-attachments/assets/f0cfc1c6-3151-4620-ba1b-be69a4318baf)

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: dan-du-car <dan.du@leidos.com>