Fix message logger plugin (#680)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

These changes are to clean up the Message Logger plugin and fix an issue
with write permissions for logs created by the plugin.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

A user wanted to use the plugin but ran into issues with starting it
because the plugin crashed on a write permissions exemption. This error
existed because the logs file already existed in the repo. When the logs
volume is created, a user is required to manually run a command to
change the folder permissions. To bypass running the command, the folder
has been removed, allowing Docker to create the folder with the correct
permissions.

Note: The file previously stored in logs/ has been moved to the
MessageLoggerPlugin/scripts/ folder. This is how other plugin scripts
are organized. A README has been added to the Message Logger Plugin to
guide users.

## How Has This Been Tested?

Tested locally on ARM64, AMD64, and with the devcontainer along with the
MessageReceiver plugin to check for correctly written logs.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [X] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.