modified:   configuration/mysql/add_v2xhub_user.bash (#681)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

Updated update-add-user.sh script to allow selecting "develop" as the
branch being used. This allows users to be created when using the
develop branch.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

Previously, user credentials were not stored correctly in the mysql
database when a developer/user chose to use the develop branch. The
script only correctly handled cases when a release was used. These
changes handle the additional case for when a develop image is used.

## How Has This Been Tested?
Tested locally with "develop" entered in the initialization script and a
second test with "develop" entered in the add_v2xhub_user script.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.