Fix MAP Plugin not broadcasting MAP in simulation (#687)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This Pull Request fixes a bug with the MAP plugin in which it only sends
a single MAP messages in simulation mode. Through local integration
testing in the Dev Container environment, it was discovered that which
the programming manual claims that multiple calls to SubscribeMessages()
should not overwrite message filters it does.

Since the PluginClockAwareClient calls attempts to subscribe to TimeSync
messages in its constructor and the MapPlugin attempts to subscribe to
ACT messages in it's constructor, the second call to Subscribe removes
the first filter. This means the plugin no longer receives time sync
messages and all calls to sleep on the carma clock will hang
indefinitely. To address this we avoid calling subscribe in the base
class and instead rely on the parent class to call subscribe.

The bug behavior can also be caught by examining plugin state

![image](https://github.com/user-attachments/assets/979c756e-4361-48f4-a5ea-39f162b220c5)
If plugin state does not include **Simulation Time Step** when in
simulation it can be concluded that the plugin message filter does not
include the TimeSync message handling of the time sync message triggers
updates to this status
<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
Fix time synchronization for MAP plugin and other impacted plugins
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
local integration testing and tested in CDASim deployment by @adev4a 
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.