Update build_v2xhub_docker_image.yml (#721)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This PR is to fix issues with the docker image build workflow when
triggered by tag events. When previously triggered by tag events we saw
the following failure
(https://github.com/usdot-fhwa-OPS/V2X-Hub/actions/runs/15619645764).

![image](https://github.com/user-attachments/assets/5505fe56-33a4-4a1c-a132-1decbcede45b)
This error occurred since the `docker/metadata-action` we use to
generate the correct image name and tag from the git hub event on git
tag generates the correct tag along with a latest tag. The default
behavior for the `docker/metadata-action` is to generate a latest tag
for any github tag event but we do not use the latest tag. To turn this
off we can use the flavor parameter to set `latest` to false to prevent
the latest tag from ever being generated by this action.
<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
Fix docker image build for tag events
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
This has been tested in the release/workzone-cp branch which includes a
tag that triggered this job
(https://github.com/usdot-fhwa-OPS/V2X-Hub/actions/runs/15620552738/job/44004066125)
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.