Add status information to FLIR Camera Driver Plugin (#723)
…d Detection Modify count

<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
This PR adds the following status counts to the FLIR Camera Driver
Plugin
- Dropped Pedestrian detections
- Modified Pedestrian detections (special cases in which we modify FLIR
Camera received detections due to confirmed bad or invalid data)

To do this we add a attribute to the SensorDetectedObject called
isModified which we set to false by default, but is intended to be set
to true when ever we need to make modifications to Sensor output due to
bad or invalid data.

In the case of the FLIR Camera, we occasionally receive detections that
have no angle (heading) but have non-zero speed. In our testing this
happens exclusively when objects become stationary. Since the SDSM
detections require both speed and heading and since speed is not
interpretable for perception with our heading, when this occurs we
simply assume the object is stationary and drop speed to 0
<!--- Describe your changes in detail -->

## Related Issue
[CDAD-149](https://usdot-carma.atlassian.net/browse/CDAD-149)
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
Prevent detection drops and provide additional diagnostic information in
V2X Hub UI.
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Unit Testing and local Integration testing.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
[V2XHUB Contributing
Guide](https://github.com/usdot-fhwa-OPS/V2X-Hub/blob/develop/Contributing.md)
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.


[CDAD-149]:
https://usdot-carma.atlassian.net/browse/CDAD-149?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ