Fix JSON Message Logger Memory leak (#781)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Add Unit tests to cover message logging functionality
Fixes memory leaks in 
- `J2735MessageFactory` Message Allocators ( This memory is pretty
obvious. Message Allocator pointers are created but never deleted.
Solved by making them smart pointers)
- 
- `TmxJ2735Codec` leaking `MessageFrameMessage` pointer

The `decode_j2735_message()` method of the `TmxJ2735EncodedMessage`
class attempts to take the UPER encoded hexadecimal `this->get_data()`
set in the message and decode it. The `MessageFrameMessage` is used as
an intermediate for conversion purposes since the data, which is a J2735
message, is often wrapped in the `MessageFrame` element. For example an
encoded MAP UPER will be wrapped in a MessageFrame element which
includes specific message IDs. For messages like MAP this is
recognizable by the first 2 bytes (00 14). This intermediate type is
then used to decode the UPER into ASN1 Compiler generated C Structs. For
SPAT this means a `MessageFrameMessage` which is a child class of
`TmxJ2735Message<MessageFrame>` will be created and an intermediate step
towards a private member variable `_decoded` which stores a
`unique_ptr<SPAT>`. Both 'MessageFrameMessage` internally has a similar
member shared_ptr member called `_j2735_data` and the `_decoded`
unique_ptr are pointing to the same underlying `*SPAT`. This means if
`MessageFrameMessage` is allocated on the stack, when it goes out of
scope its `shared_ptr` will call the registered deleter which ultimately
calls `ASN_STRUCT_FREE` (see `j2735_destroy` in `SaeJ2735Traits.hpp` and
`TmxJ2735Message` class ). This will delete the same `*SPAT` that the
`TmxJ2735EncodedMessage` is storing as a `unique_ptr` in its `_decoded`
member.

The ideal solution would be to resolve the issue that both a unique and
shared pointer are pointer to the same underlying memory since that is a
violation of the implied ownership of the memory. (can't be owned by the
unique_ptr and the shared_ptr), but a more short term safe solution here
is to ensure that the two objects that share this memory also share the
same object scope which is what this PR does.

<!--- Describe your changes in detail -->

## Related GitHub Issue

<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
[VH-1457](https://usdot-carma.atlassian.net/browse/VH-1457)
<!-- e.g. CAR-123 -->

## Motivation and Context
Fix Memory leaks in code
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Running Unit Tests
Running valgrind on JSON Message Logger Unit Test
Running valgrind on Telematics Bridge Plugin Unit test
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.


[VH-1457]:
https://usdot-carma.atlassian.net/browse/VH-1457?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ