Updated upload file path to be in download folder to resolve permission issues (#792)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
When using the upload functionality on the web UI and selecting the MAP
type file uploads would fail, seemingly related to permission issues
with folders under `/var/www/plugins/`. This PR updates the default
download path for MAP files to be `/var/www/download/MAP` which is under
the default path for all downloads and does not have this permission
issue.

This PR included fixes to ensure back slashes were correctly added to
the destination file path as well as fixing issues related to `Invalid
cross-device link` errors when attempting to copy downloaded files from
the `/var/www/download/` directory to the default MAP directory of
`/var/www/download/MAP`.
According to google AI this error can occur when using docker volumes
because :
```
The boost::filesystem::copy_file error "Invalid cross-device link" (EXDEV) in a Docker environment occurs because you are attempting a file operation that requires the source and destination paths to be on the same filesystem/mount point, but they are located on different ones. This is a common issue in Docker when using multiple volumes or bind mounts, as each mount point is treated as a separate device by the operating system within the container
```
To resolve the issue I simply ensured that the entire download folder
was the volume which would prevent copy operations from not being on the
same file system. I think we may be able to make this a named volume
eventually which is not directly accessible on the host machine. This
would preserve state between shutdowns but also prevent unintentional
file inclusion and provide clear single way to include files in V2X Hub
<!--- Describe your changes in detail -->

## Related GitHub Issue
NA
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
VH-1468
<!-- e.g. CAR-123 -->

## Motivation and Context
Fix MAP file web UI upload functionality
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
1) Open web UI
2) Click file upload 
3) Select MAP type
4) Select MAP file
5) Upload
6) Configure MAP Plugin to use this file
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.